requirements.txt
accelerate==1.7.0
aif360==0.6.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
audioop-lts==0.2.1
av==13.1.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blis==1.3.0
catalogue==2.0.10
chardet==5.2.0
charset-normalizer==3.4.3
choreographer==1.0.9
click==8.2.1
cloudpathlib==0.22.0
cloup==3.0.7
comm==0.2.2
confection==0.1.5
cymem==2.0.11
Cython==3.1.2
dataclasses-json==0.6.7
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.33.1
dill==0.3.8
distro==1.9.0
executing==2.2.0
faiss-cpu==1.11.0.post1
fastjsonschema==2.21.1
filelock==3.18.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
geopandas==1.1.0
glcontext==3.0.0
h11==0.16.0
hf-xet==1.1.3
hiddenlayer==0.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.32.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
isosurfaces==0.1.2
jedi==0.19.2
jiter==0.10.0
json5==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
jupytext==1.17.2
kaleido==1.0.0
kiwisolver==1.4.8
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-deepseek==0.1.4
langchain-huggingface==0.2.0
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langcodes==3.5.0
langsmith==0.3.45
language_data==1.3.0
lap==0.5.12
lightgbm==4.6.0
logistro==1.1.0
Mako==1.3.10
mapbox_earcut==1.0.3
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==3.0.0
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.3
moderngl==5.12.0
moderngl-window==3.1.1
moviepy==2.2.1
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.16
munkres==1.1.4
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==2.0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.3
notebook_shim==0.2.4
numpy==2.3.0
openai==1.97.1
opencv-python==4.11.0.86
openTSNE==1.0.2
orjson==3.10.18
osmnx==2.0.6
overrides==7.7.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pdoc3==0.11.6
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
plotly==6.2.0
preshed==3.0.10
proglog==0.1.12
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pydantic==2.11.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyglet==2.1.6
pyglm==2.8.2
Pygments==2.19.1
pyogrio==0.11.0
pypdf==5.9.0
pyproj==3.7.1
python-dotenv==1.1.1
python-json-logger==3.3.0
PyYAML>=6.0
pyzmq==26.4.0
qwen-vl-utils==0.0.11
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.7.0
screeninfo==0.8.1
Send2Trash==1.8.3
sentence-transformers==4.1.0
setuptools==72.1.0
shapely==2.1.1
shellingham==1.5.4
simplejson==3.20.1
skia-pathops==0.8.0.post2
smart_open==7.3.1
sniffio==1.3.1
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
srt==3.5.3
stack-data==0.6.3
svgelements==1.9.6
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
thinc==8.3.6
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
torch==2.7.1
torchsummary==1.5.1
torchvision==0.22.1
tornado==6.5.1
traitlets==5.14.3
transformers==4.52.4
typer==0.19.1
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing-inspection==0.4.1
ultralytics==8.3.159
ultralytics-thop==2.0.14
uri-template==1.3.0
urllib3==2.4.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0

